V0.8.1
Created second copy of the project to separate manual control from automatic, in order to be able to test them without removing the angle wires.

Also replaced a number of blocking assignments with non blocking assignments.

Angle 3 still does not work properly, I believe it is a hardware issue,

Angle 4 still has issues, I may consider replacing it with a static pen but don't know if it will be able to reach the whole maze, might need to shrink each square to fit them all in.